Add connection to MongoDB with autopopulation, but lacks questionId compatability and separation of autopopulation
