+++ questionbank/src/main/java/com/example/questionbank/commons/QuestionNotFoundException.java
@@ -2,7 +2,7 @@ package com.example.questionbank.commons;
 
 public class QuestionNotFoundException extends RuntimeException {
 
-    public QuestionNotFoundException(Long id) {
+    public QuestionNotFoundException(String id) {
         super("Could not find question " + id);
     }
 }
