+++ questionbank/src/main/java/com/example/questionbank/model/Question.java
@@ -1,37 +1,37 @@
 package com.example.questionbank.model;
 
 import java.util.List;
-import java.util.Objects;
 
-import jakarta.persistence.Column;
-import jakarta.persistence.ElementCollection;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
+import org.springframework.data.annotation.Id;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.AllArgsConstructor;
+import org.springframework.data.mongodb.core.index.Indexed;
+import org.springframework.data.mongodb.core.mapping.Document;
 
 
-@Entity
 @Data // Lombok generates getters, setters, equals, hashCode and toString
 @NoArgsConstructor // Lombok generates a no-argument constructor
 @AllArgsConstructor // Lombok generates a constructor with all fields
+@Document
 public class Question {
 
-    private @Id
-    @GeneratedValue Long id;
+    @Id
+    private String id; // MongoDB id
+
+    @Indexed(unique = true)
+    private Long questionId; // Numeric question ID, unique and indexed
+
+    @Indexed(unique=true)
     private String title;
 
-    @Column(length = 1000) // Adjust length to fit description needs
     private String description;
 
-    @ElementCollection // Indicates that this is a collection of basic types
-    private List<String> categories; // A list of categories
+    private List<String> categories;
 
     private String complexity;
 
-    // Constructor without id field
+    // Constructor without MongoDB Vid field
     public Question(String title, String description, List<String> categories, String complexity) {
 
         this.title = title;
