+++ questionbank/src/main/java/com/example/questionbank/repository/QuestionRepository.java
@@ -1,8 +1,8 @@
 package com.example.questionbank.repository;
 
-import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.mongodb.repository.MongoRepository;
 import com.example.questionbank.model.Question;
 
-public interface QuestionRepository extends JpaRepository<Question, Long> {
-
+public interface QuestionRepository extends MongoRepository<Question, String> {
+    public Question findQuestionByTitle(String title); // Example Spring Data Commons derived method for queries
 }
