+++ questionbank/src/main/resources/application.properties
@@ -1,1 +1,8 @@
 spring.application.name=questionbank
+spring.data.mongodb.database=peerprep
+spring.data.mongodb.username=jaime
+spring.data.mongodb.password=cs3219peerprepg08
+spring.data.mongodb.port=27017
+spring.data.mongodb.host=localhost
+spring.data.mongodb.auto-index-creation=true
+spring.data.mongodb.uri=mongodb+srv://jaime:cs3219peerprepg08@peerprep.ycbdz.mongodb.net/?retryWrites=true&w=majority&appName=PeerPrep
